Route all user profiles to /username
For simplicity, all profiles go through /{username}. This allows for easier navigation and search for users. The page uses conditionals to serve specific content based on whether the user is signed in or not (logout, follow, etc.).
