+++ backend/urls.py
@@ -10,7 +10,7 @@ urlpatterns = [
     path('api/', include([
         path('auth/', include("pages.authentication.urls", namespace="authentication")),
         path('discover/', GetUsersView.as_view(), name="get_users"),
-        path('create/', CreatePostView.as_view(), name='create_post'),
+        path('create-post/', CreatePostView.as_view(), name='create_post'),
         path('musician/<uuid:user_id>/', MusicianDetailView.as_view(), name='musician-detail'),
         path('instruments/', InstrumentListView.as_view(), name='instrument-list'),
         path('genres/', GenreListView.as_view(), name='genre-list'),
